Begin separate KB key transforms test.
